Refs #785: now when donor is unspecified, choose donor from the same segment first.
git-svn-id: https://devel.codership.com/svn/galera/branches/3.x@3259 9148e567-b586-4176-b9b8-4e36cb4ec61b

+++ gcs/src/gcs_recv_msg.h
@@ -16,9 +16,9 @@
 typedef struct gcs_recv_msg
 {
     void*          buf;
-    long           buf_len;
-    long           size;
-    long           sender_idx;
+    int            buf_len;
+    int            size;
+    int            sender_idx;
     gcs_msg_type_t type;
 }
 gcs_recv_msg_t;
